https://issues.jboss.org/browse/SWITCHYARD-154 Add support for creating and registering a Transformer via a TransformModel config instance
