+++ transform-smooks/src/test/resources/xml/order.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<order xmlns="urn:switchyard-quickstarts:transform-smooks:1.0">
+<ord:order xmlns:ord="urn:switchyard-quickstarts:transform-smooks:1.0">
     <orderId>PO-19838-XYZ</orderId>
     <itemId>BUTTER</itemId>
     <quantity>200</quantity>
-</order>
+</ord:order>
